Several bugfixes in restore script
After my initial test I found a couple of bugs that really should have
found in Vagrant testing and review.

Regarding the mv / cp of the current /var/lib/mysql directory, this was
an issue because that directory is a mount point so it is unable to be
moved. We can copy and delete instead.